not truncating table
this is done to workaround the bug in which the default stream is not available
immediately when first creating a table. In that case the table used to be
always created fresh and the stream was not available after each recreation. Not
truncating is fine because we are using BQ cdc enabled tables in which inserting
duplicate UPSERT entries should be fine.