fix: dangling open transaction for cdc-only feat: add nats connector (for cdc events)
For the first part, fix START_REPLICATION cannot run inside a
transaction block for CDC only, which is caused by
PostgresSource::new opening a transaction, but never closing it
for CDC only. This can be remedied by committing that transaction
and offering an explicit start_transaction function to be called
in cases of not CDC only.

For the 2nd part, introduce a NATS connector (WIP,
this should probably create or get a stream in Jetstream
for persistence and needs some configuration options)
so we can propagate out CDC events onto a message broker

Refs: #80